add bottom nav bar
