+++ front-end/src/components/common/Logo.jsx
@@ -1,16 +1,17 @@
-import { useNavigate } from "react-router-dom"
+import { useNavigate, useLocation } from "react-router-dom"
 
 const Logo = () => {
   const navigate = useNavigate();
+  const location = useLocation();
   const handleClick = (evt) => {
     evt.stopPropagation()
-    navigate("/")
-    }
+    navigate("/", { state: { from: location.pathname } })
+  }
 
   return (
     <>
     {/* Reference: <a href="https://www.flaticon.com/free-icons/map" title="map icons">Map icons created by Freepik - Flaticon</a> */}
-    <img className="w-9 hover:cursor-pointer" src="/maplogo.png" alt="maplogo" onClick={handleClick}/>
+    <img className="w-7 h-7 hover:cursor-pointer" src="/maplogo.png" alt="maplogo" onClick={handleClick}/>
     </>
   )
 }
