+++ front-end/src/components/common/pageLink.jsx
@@ -1,13 +1,13 @@
 import { Link } from 'react-router-dom'
-
+import { useLocation } from 'react-router-dom';
 const PageLink = (props) => {
   // attributes: to, value
-  // optional: from
+  const location = useLocation();
   return(
     <div className='w-full text-center py-2 underline'>
       <Link 
         to={props.to}
-        state={{from: props.from}}>
+        state={{from: location.pathname}}>
         {props.value}
       </Link>
     </div>
