+++ front-end/src/components/common/rightBtn.jsx
@@ -1,7 +1,7 @@
 const RightBtn = (props) => {
   return (
     <>
-    <img className="w-9 hover:cursor-pointer" src="/rightbtn.png" alt="rightbtn" onClick={props.handleRightClick}/>
+    <img className="w-7 h-7 hover:cursor-pointer" src="/rightbtn.png" alt="rightbtn" onClick={props.handleRightClick}/>
     </>
   )
 }
