+++ front-end/src/components/popup/popupContent.jsx
@@ -6,7 +6,7 @@ function PopupContent(props){
   return (
     // dark background
     <div 
-      className="popupBackground fixed top-0 left-0 z-50 w-full h-full bg-black bg-opacity-50 flex items-center justify-center" 
+      className="popupBackground fixed top-0 left-0 z-[1700] w-full h-full bg-black bg-opacity-50 flex items-center justify-center" 
       onClick={props.handleClick}>
       {/* white popup container: */}
       <div className="bg-white rounded-lg shadow-xl w-[80%]">
