+++ front-end/src/pages/Error/Error.jsx
@@ -1,12 +1,15 @@
 import PageLink from "../../components/common/pageLink"
-import { useLocation } from "react-router-dom"
+import BottomNavBar from "../../components/common/bottomNavBar";
 const Error = () => {
-  const location = useLocation();
   return(
     <>
-      <h1 className="w-full text-center ">404 Not Found</h1>
-      <PageLink to="/" value="Go back to main page"/>
-      <PageLink to="/info" value="test info" from={location.pathname}/>
+      <div className="h-screen flex">
+        <div className="mx-auto mt-[20vh]">
+          <h1 className="w-full text-center">404 Not Found</h1>
+          <PageLink to="/" value="Go back to main page"/>
+        </div>
+      </div>
+      <BottomNavBar />
     </>
   )
 }
