+++ front-end/src/pages/InfoDetail/InfoDetail.jsx
@@ -1,3 +1,7 @@
+import NavBar from "../../components/common/navBar"
+import LeftBtn from "../../components/common/leftBtn"
+import BottomNavBar from "../../components/common/bottomNavBar"
+
 const InfoDetail = (props) => {
   // parameters: pic, name, subtitle(string array), parag
   const dummyPic = "https://picsum.photos/200"
@@ -10,23 +14,29 @@ const InfoDetail = (props) => {
 
   return (
     <>
-      <div className="max-h-[80vh] max-w-full m-auto flex">
-        <img className="object-contain" src={props.pic ?? dummyPic} alt="picture" />
+      <div className="flex flex-col">
+        <NavBar>
+          <LeftBtn />
+        </NavBar>
+        <div className="max-h-[80vh] max-w-full mx-auto flex">
+          <img className="object-contain" src={props.pic ?? dummyPic} alt="picture" />
+        </div>
+        <div className="w-[80%] mb-[10%] mx-auto">
+          <div className="mt-2">
+            <h2>{props.name ?? dummyName}</h2>
+          </div>
+          <div className="mt-2">
+            {props?.subsitle ?
+              props?.subtitle?.map((subtitle, i) => <p key={i}>{subtitle}</p>) :
+              dummySubtitle?.map((subtitle, i) => <p key={i}>{subtitle}</p>)
+            }
+          </div>
+          <div className="mt-4">
+            <p>{props.parag ?? dummyparag}</p>
+          </div>
+        </div>
       </div>
-      <div className="w-[80%] mb-[10%] mx-auto">
-        <div className="mt-2">
-          <h2>{props.name ?? dummyName}</h2>
-        </div>
-        <div className="mt-2">
-          {props?.subsitle ?
-            props?.subtitle?.map((subtitle, i) => <p key={i}>{subtitle}</p>) :
-            dummySubtitle?.map((subtitle, i) => <p key={i}>{subtitle}</p>)
-          }
-        </div>
-        <div className="mt-4">
-          <p>{props.parag ?? dummyparag}</p>
-        </div>
-      </div>
+      <BottomNavBar relative="1"/>
     </>
   )
 }
